fix(backend): align runtime projection contract boundary
