+++ src/backend/src/agentclaw/community/core/skill_center/services/local_skill_upload_service.py
@@ -55,7 +55,7 @@ from agentclaw.community.core.skills_pool.edit_guard import (
     SkillsPoolEditRollbackError,
 )
 from agentclaw.community.core.skills_pool.types import BotSkillLayoutScope
-from agentclaw.community.api.bot_runtime_projection_reconciler import (
+from agentclaw.community.core.skill_center.runtime_projection_contract import (
     BotRuntimeProjectionReconcilerProtocol,
 )
